scheduling/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0